chore: mock

+++ src/services/transaction.ts
@@ -3,11 +3,16 @@
 import { Filters } from '@/@types/filters';
 import { Transaction } from '@/@types/transaction';
 import { convertToSearchString, onlyKeysWithTruthyValueAndConvertToString } from '@/utils/object';
+import { MOCK_TRANSACTIONS } from './mockTransactions';
 
 const BASE_URL = `${process.env.REACT_API_URL}/transactions`;
 
 export const createTransaction = async (transaction: Omit<Transaction, 'id'>) => {
   'use server';
+  if (process.env.NODE_ENV === 'production') {
+    return MOCK_TRANSACTIONS[0];
+  }
+
   const response = await fetch(BASE_URL, {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
@@ -19,6 +24,11 @@ export const createTransaction = async (transaction: Omit<Transaction, 'id'>) =>
 
 export const getTransactionsByUser = async (author: string, filters: Filters<Transaction> = {}) => {
   'use server';
+  if (process.env.NODE_ENV === 'production') {
+    if (author === 'e00a') {
+      return MOCK_TRANSACTIONS;
+    }
+  }
   const searchParams = convertToSearchString(onlyKeysWithTruthyValueAndConvertToString({ author, ...filters }));
   const response = await fetch(`${BASE_URL}?${searchParams}`);
   return response.json() as Promise<Transaction[]>;
@@ -27,6 +37,10 @@ export const getTransactionsByUser = async (author: string, filters: Filters<Tra
 export const getTransactionById = async (id: string) => {
   'use server';
 
+  if (process.env.NODE_ENV === 'production') {
+    return MOCK_TRANSACTIONS.find((transaction) => transaction.id === id) || MOCK_TRANSACTIONS[0];
+  }
+
   const response = await fetch(`${BASE_URL}/${id}`);
   const result: Transaction = await response.json();
   return result;
@@ -34,6 +48,10 @@ export const getTransactionById = async (id: string) => {
 
 export const updateTransaction = async (transaction: Transaction) => {
   'use server';
+
+  if (process.env.NODE_ENV === 'production') {
+    return transaction;
+  }
 
   const response = await fetch(`${BASE_URL}/${transaction.id}`, {
     method: 'PATCH',
@@ -46,6 +64,10 @@ export const updateTransaction = async (transaction: Transaction) => {
 
 export const deleteTransaction = async (id: string) => {
   'use server';
+
+  if (process.env.NODE_ENV === 'production') {
+    return MOCK_TRANSACTIONS.find((transaction) => transaction.id === id);
+  }
 
   const response = await fetch(`${BASE_URL}/${id}`, {
     method: 'DELETE',
